% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stics_exe.R
\name{add_stics_exe}
\alias{add_stics_exe}
\title{Adding a new Stics model executable in JavaStics}
\usage{
add_stics_exe(javastics_path, stics_exe, exe)
}
\arguments{
\item{javastics_path}{JavaStics installation root folder}

\item{stics_exe}{Stics executable name/identifier (not executable file)}

\item{exe}{Stics executable name or path}
}
\description{
Add a new executable of the Stics model to use in JavaStics, and set it as
the one to use by default. Please refer to \code{set_stics_exe()} to change the stics
executable used by JavaStics, and \code{list_stics_exe()} to list all available executables.
}
\details{
The executable file may be in the current JavaStics bin folder. If not,
the \code{exe} argument is considered as a path and the executable will be copied in
the bin directory.
}
\examples{
\dontrun{
 add_stics_exe("/path/to/JavaSTICS/dir","model_name","model_exe_name")
}

}
