% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_out_var.R
\name{all_out_var}
\alias{all_out_var}
\title{Return all possible STICS outputs for var.mod}
\usage{
all_out_var()
}
\description{
Helper function to print the list of all possible variables to set as output
from the STICS model.
}
\examples{
library(sticRs)
All_vars= all_out_var()

}
\seealso{
\code{\link{set_out_var}}
}
