% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_java_workspace.R
\name{check_java_workspace}
\alias{check_java_workspace}
\title{Checking and getting JavaStics workspace path}
\usage{
check_java_workspace(javastics_path, javastics_workspace_path = NULL)
}
\arguments{
\item{javastics_path}{Path of JavaStics installation directory}

\item{javastics_workspace_path}{An absolute or relative path (to JavaStics path)
of a workspace (Optional)}
}
\value{
An absolute javastics workspace path
}
\description{
Looking in JavaStics for a workspace path in JavasStics preferences
or producing a full path to a workspace located in JavaStics root directory
or validating an external absolute path. And also checking if the path is a
valid JavaStics workspace
}
\keyword{internal}
