% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_output_files.R
\name{check_output_files}
\alias{check_output_files}
\title{Checking model outputs files}
\usage{
check_output_files(run_dir, usm_name = NULL)
}
\arguments{
\item{run_dir}{Directory name or path (absolute) of a working directory,
or root directory of working directories}

\item{usm_name}{Usm name, to be set if it is the name of a sub-directory of the given argument run_dir}
}
\value{
A list with fields $error (logical, TRUE if any missing file, FALSE otherwise), $missing (missing files vector)
}
\description{
This function checks if the model outputs files are correctly produced
}
\examples{
\dontrun{
check_output_files("/home/username/Work/SticsInputsDir")
check_output_files("/home/username/Work/SticsInputsRootDir","usmDir")
}

}
\keyword{internal}
