% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_2usms.R
\name{compare_2usms}
\alias{compare_2usms}
\title{Comparing Stics daily outputs of 2 directories}
\usage{
compare_2usms(first_dir, second_dir, in_usms_list = c())
}
\arguments{
\item{first_dir}{A first output directory path}

\item{second_dir}{A second output directory path}

\item{in_usms_list}{Usms selection list inside usms from usms.xml file (optional)}
}
\value{
A list with usms names, usms correlations for variables, and bad NULL and NA status !!!
}
\description{
For common usm, or a given usm list, and for common variables correlations are calculated
}
\examples{
out <- compare_2usms("out_dir1","out_dir2")
out <- compare_2usms("out_dir1","out_dir2",c("ble","mais"))
}
