% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_doy_cumul.R
\name{compute_doy_cumul}
\alias{compute_doy_cumul}
\title{Calculating cumulative day of year from days of year over 2 years}
\usage{
compute_doy_cumul(doys_vec, years_vec)
}
\arguments{
\item{doys_vec}{vector of days of year}

\item{years_vec}{vector of years corresponding to each day of year from doys_vec}
}
\value{
A vector of cumulative days of year over a 2 years period
}
\description{
For a vector of days of year and a vector of corresponding year
for each one, calculation of cumulative days of years
}
\examples{
compute_doy_cumul(c(350,360,10,20,30),c(1990,1990,1991,1991,1991))
}
