% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Del_spe_col.R
\name{Del_spe_col}
\alias{Del_spe_col}
\title{Erase special characters from STICS outputs or observations}
\usage{
Del_spe_col(x)
}
\arguments{
\item{x}{A data.frame}
}
\value{
A data.frame with R-ready column names.
}
\description{
Helper function to replace the first parenthesis by a
"_" and erase the second. For exemple, lai(n) becomes lai_n.
}
\examples{
\dontrun{
library(sticRs) ; library(data.table)
a= data.table(`lai(n)`= c(1:8))
Del_spe_col(a)
}
}
