% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_stics_exe.R
\name{detect_stics_exe}
\alias{detect_stics_exe}
\title{Detecting Stics executables files according to the host OS}
\usage{
detect_stics_exe(dir_path)
}
\arguments{
\item{dir_path}{JavaStics bin directory path
(or another one containing Stics executable files)}
}
\value{
A data.frame with columns: path (executable path),
exe (executable file name), version (from the executable)
}
\description{
Detecting Stics executables files according to the host OS
}
\keyword{internal}
