% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_output.R
\name{eval_output}
\alias{eval_output}
\title{Compare STICS simulation outputs (mod_s) to observations (*.obs)}
\usage{
eval_output(dirpath = getwd(), obs_name = NULL, mixed = NULL)
}
\arguments{
\item{dirpath}{Directory path}

\item{obs_name}{A vector of observation file name(s). Optional, see details.}

\item{mixed}{(optional) Is the simulation made on mixed species (boolean)}
}
\value{
A data.frame (sole crop) or a list of two data.frames (mixed crops) with
        simulated and observed data. Simulated and measured data have the \code{_sim}
        and \code{_meas} suffix respectively.
}
\description{
Compare STICS simulation outputs to observations for sole or mixed crops.
}
\details{
If \code{obs_name} is not provided, the function try to guess it using the
         built-in algorithm from \code{\link{read_obs}}. Idem for the \code{mixed}
         argument. See documentation for more details.
}
\examples{
\dontrun{
library(sticRs)
Table_compare= eval_output()
}

}
\seealso{
\code{\link{read_output}} and \code{\link{read_obs}}
}
