% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_java_model.R
\name{exist_java_model}
\alias{exist_java_model}
\title{Getting existing status of a Stics model name}
\usage{
exist_java_model(javastics_path, java_model_tag)
}
\arguments{
\item{javastics_path}{JavaStics installation root folder}

\item{java_model_tag}{Model name in JavaStics configuration}
}
\value{
Existing status, logical
}
\description{
Finding in JavaStics preference.xml file in the given JavaStics path,
if the given model name exists
}
\examples{
\dontrun{
exist_java_model("/home/plecharpent/Work/JavaSTICS-v131-stics-v841/example","stics_name")
}

}
\keyword{internal}
