% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_exe_utilities.R
\name{exist_stics_exe}
\alias{exist_stics_exe}
\title{Check if a stics executable is available}
\usage{
exist_stics_exe(javastics, stics_exe)
}
\arguments{
\item{javastics}{JavaStics installation root folder}

\item{stics_exe}{Stics executable name (see details)}
}
\value{
Existing status, logical
}
\description{
Checks if a stics model executable is available in JavaStics (in the
"preference.xml" file).
}
\details{
The current model executable available in JavaStics can be listed using
\code{list_stics_exe()}.
}
\examples{
\dontrun{
exist_stics_exe("path/to/JavaSTICS-v131-stics-v841","stics_name")
}

}
\keyword{internal}
