% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists_java_pref.R
\name{exists_java_pref}
\alias{exists_java_pref}
\title{Evaluate if model preferences hav been set}
\usage{
exists_java_pref(javastics_path)
}
\arguments{
\item{javastics_path}{JavaStics installation root folder}
}
\value{
logical value, TRUE if file exists, FALSE otherwise
}
\description{
Testing if preferences.xml file exist in JavaSTICS installation folder
}
\examples{
\dontrun{
exists_pref <- exists_java_pref("/home/plecharpent/Work/JavaSTICS-v131-stics-v841")
}

}
