% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_exe_utilities.R
\name{exists_javastics_pref}
\alias{exists_javastics_pref}
\title{Evaluate if model preferences have been set}
\usage{
exists_javastics_pref(javastics_path)
}
\arguments{
\item{javastics_path}{JavaStics installation folder}
}
\value{
logical value, TRUE if file exists, FALSE otherwise
}
\description{
Testing if preferences.xml file exist in
JavaSTICS installation folder
}
\examples{
\dontrun{
exists_pref <- exists_javastics_pref("path/to/JavaSTICS-v131-stics-v841")
}

}
\keyword{internal}
