% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_out_var.R
\name{find_STICS_var}
\alias{find_STICS_var}
\title{Find output variable names for STICS}
\usage{
find_STICS_var(Var)
}
\arguments{
\item{Var}{Character vector with a (partial) STICS output variable name}
}
\description{
Helper function that return the name used as input
for the STICS model with a partial match.
}
\details{
The function understand \code{\link[base]{regex}} as input.
}
\examples{
library(sticRs)
find_STICS_var("lai")

}
\seealso{
\code{\link{all_out_var}}
}
