% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_usms_dirs.R
\name{gen_usms_dirs}
\alias{gen_usms_dirs}
\title{Generating one or several usms directories from a javastics workspace
content}
\usage{
gen_usms_dirs(javastics_path, javastics_workspace_path = NULL,
  target_path = NULL, usms_list = c(), display = FALSE,
  flag_dirPerUsm = TRUE)
}
\arguments{
\item{javastics_path}{Path of JavaStics installation directory}

\item{javastics_workspace_path}{Path of a JavaStics workspace (Optional)}

\item{target_path}{The path of the directory where to create usms directories (Optional),
if not provided the JavaStics workspace will be used as root}

\item{usms_list}{List of usms to run (Optional)}

\item{display}{Logical value for displaying (TRUE) ot not (FALSE) usm name}

\item{flag_dirPerUsm}{logical, TRUE if one want to create one directory per USM,
FALSE if USM files are generated in the target_path (only useful for usms_list of size one)}
}
\value{
A list of created directories with Stics input files inside or
NULL if any problem about the JavaStics workspace or JavaStics directory
}
\description{
TODO
}
\examples{
TODO
}
