% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_usms_xml2txt.R
\name{gen_usms_xml2txt}
\alias{gen_usms_xml2txt}
\title{Generating one or several usms directories from a javastics workspace
content}
\usage{
gen_usms_xml2txt(
  javastics_path,
  javastics_workspace_path = NULL,
  target_path = NULL,
  usms_list = c(),
  display = FALSE,
  dir_per_usm_flag = TRUE
)
}
\arguments{
\item{javastics_path}{Path of JavaStics installation directory}

\item{javastics_workspace_path}{Path of a JavaStics workspace (Optional), if not
provided the current workspace stored in JavaStics preferences will be used.}

\item{target_path}{The path of the directory where to create usms directories (Optional),
if not provided the JavaStics workspace will be used as root}

\item{usms_list}{List of usms to generate (Optional). If not provided, all
usms contained in javastics_workspace_path/usms.xml file will be generated.}

\item{display}{Logical value for displaying (TRUE) ot not (FALSE) usm name}

\item{dir_per_usm_flag}{logical, TRUE if one want to create one directory per USM,
FALSE if USM files are generated in the target_path (only useful for usms_list of size one)}
}
\value{
A list with named elements:
usms_paths : created directories names (Stics input files),
files_path : generated files list paths (in JavaStics workspace origin),
copy_status : logical value vector, indicating if all files have been generated
obs_copy_status : logical value vector, indicating if observaion files have been
successfully copied in usms directories
}
\description{
The function creates sets of input files for one or multiple usms
from usms data stored in a JavaStics workspace. For multiple usms, sets will
be generated into individual folders named with usm names.. Observations
files will be also copied if they are named \verb{[usm_name].obs}
For one usm, files will be generated either in the workspace directory
or in a subdirectory.
}
\examples{
\dontrun{
javastics <- "/path/to/javastics"
javastics_workspace <- "/path/to/workspace"

# For all usms
gen_usms_xml2txt(javastics, javastics_workspace)

# For an usms list
usm_list <- c("usm1", "usm2")
gen_usms_xml2txt(javastics, javastics_workspace, usm_list)

# For one usm
gen_usms_xml2txt(javastics, javastics_workspace, dir_per_usm_flag=F, "usm1")

}
}
