% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cores_nb.R
\name{get_cores_nb}
\alias{get_cores_nb}
\title{Getting the available number of cores for parallel calculations}
\usage{
get_cores_nb(parallel = FALSE, required_nb = NA, ...)
}
\arguments{
\item{parallel}{Logical for performing parallel loop (TRUE) or not (FALSE)}

\item{required_nb}{Wanted number of cores}

\item{...}{To pass additional arguments (i.e. cores_nb, fake machine cores)}
}
\value{
Available cores number tu use
}
\description{
Getting the available number of cores for parallel calculations
}
\examples{
\dontrun{

get_cores_nb()

get_cores_nb(parallel = TRUE)

get_cores_nb(parallel = TRUE, required_nb = 4)
}

}
\keyword{internal}
