% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_java_model.R
\name{get_java_model}
\alias{get_java_model}
\title{Getting current Stics instance in use}
\usage{
get_java_model(javastics_path)
}
\arguments{
\item{javastics_path}{JavaStics installation root folder}
}
\value{
Current Stics model name used in JavaStics configuration
}
\description{
Extracting the last used model from preferences file
}
\examples{
\dontrun{
current_model <- get_java_model("/home/plecharpent/Work/JavaSTICS-v131-stics-v841")
}

}
