% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_java_models.R
\name{get_java_models}
\alias{get_java_models}
\title{Getting models list from a JavaStics configuration}
\usage{
get_java_models(javastics_path)
}
\arguments{
\item{javastics_path}{JavaStics installation root folder}
}
\value{
A list with model name list ($tag), and model executables names ($exe)
}
\description{
Extracting Stics model names and executables file names from preferences file
}
\examples{
models <- get_java_models("/home/plecharpent/Work/JavaSTICS-v131-stics-v841")

}
