% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_java_models_nb.R
\name{get_java_models_nb}
\alias{get_java_models_nb}
\title{Getting the number of Stics executable registered in JavaStics}
\usage{
get_java_models_nb(javastics_path, check = FALSE)
}
\arguments{
\item{javastics_path}{JavaStics installation root folder}

\item{check}{Logical for executing or not a check on executables files}
}
\value{
a number of registered models in JavaStics preferences,
and corresponding exiseing executable file
}
\description{
Getting the number of Stics executable registered in JavaStics
}
\keyword{internal}
