% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_java_models_version.R
\name{get_java_models_version}
\alias{get_java_models_version}
\title{Getting Stics versions list of executable in a JavaStics directory}
\usage{
get_java_models_version(javastics_path, exe_tags = NULL)
}
\arguments{
\item{javastics_path}{JavaStics installation root folder}

\item{exe_tags}{Versions tags vector}
}
\value{
A named list ( with model tags ) of versions strings
(got from the model executable)
}
\description{
Getting Stics versions list of executable in a JavaStics directory
}
