% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_java_workspace.R
\name{get_java_workspace}
\alias{get_java_workspace}
\title{Getting current workspace in JavaStics preferences configuration}
\usage{
get_java_workspace(javastics_path)
}
\arguments{
\item{javastics_path}{JavaStics installation root folder}
}
\description{
Getting current JavaStics working directory,
if not any setting to \code{example} directory
}
\examples{
\dontrun{
get_java_wd("/path/to/JavaStics/directory")
}

}
\keyword{internal}
