% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_javastics.R
\name{get_javastics_cmd}
\alias{get_javastics_cmd}
\title{Get JavaSTICS command line executable name}
\usage{
get_javastics_cmd(javastics)
}
\arguments{
\item{javastics}{JavaSTICS installation root folder}
}
\value{
An executable name
}
\description{
From a JavaSTICS path, searching for the JavaSTICS command
}
\details{
If JavaStics version < 1.5, returns \code{JavaSticsCmd.exe}
otherwise, returns \code{JavaStics.exe}
}
\examples{
\dontrun{
 get_javastics_cmd("path/to/JavaSTICS")
}

}
\keyword{internal}
