% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_exe.R
\name{get_model_exe}
\alias{get_model_exe}
\title{Getting Stics model executable name}
\usage{
get_model_exe(javastics_path, model_name = character(0))
}
\arguments{
\item{javastics_path}{JavaStics installation root folder}

\item{model_name}{Model name in model name list}
}
\value{
Stics model executable name (current, matching given model name)
}
\description{
Extracting the current used model executable name from preferences file

or corresponding one to the given model_name
}
\examples{
Getting the current model executable
model_exe <- get_model_exe("/home/plecharpent/Work/JavaSTICS-v131-stics-v841")
Getting model executable correponding to the given "model_name"
model_exe <- get_model_exe("/home/plecharpent/Work/JavaSTICS-v131-stics-v841","model_name")

}
