% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stics_models_exe.R
\name{get_stics_models_exe}
\alias{get_stics_models_exe}
\title{Stics executable version}
\usage{
get_stics_models_exe(javastics_path, stics_exe = NULL)
}
\arguments{
\item{javastics_path}{JavaStics installation root folder}

\item{stics_exe}{Stics executable name (see details)}
}
\value{
A named list with stics executables from the model executable
}
\description{
Return the model version of each stics executable available in
JavaStics (using --version).
}
\details{
The \code{stics_exe} is \strong{not} the name of the executable file, but the
identification name. Please use \code{list_stics_exe()} to list all available executables,
and \code{set_stics_exe()} to add a new one. The identification names can be retreived using
\code{names(list_stics_exe(javastics_path)$stics_list)}
}
\keyword{internal}
