% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_exe_utilities.R
\name{get_version_number}
\alias{get_version_number}
\title{Get STICS executable version number, or string, or hash}
\usage{
get_version_number(exe_path, numeric = TRUE)
}
\arguments{
\item{exe_path}{executable file path}

\item{numeric}{Logical, \code{TRUE} (default) to return a numeric version as a
semver class vector, or \code{FALSE} to return a character string with the version}
}
\value{
a smvr class vector if \code{numeric = TRUE} (default),
or a character string if \code{numeric = FALSE}, with an attribute "date" if
the version date is provided in the system command output.
Or a hash string if the system command output contains a hash.
}
\description{
Get STICS executable version number, or string, or hash
}
\examples{
\dontrun{
get_version_number("path/to/stics_exe")
get_version_number("path/to/stics_exe", numeric = FALSE)
}

}
