% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_output.R
\name{identical_vals}
\alias{identical_vals}
\title{Identical}
\usage{
identical_vals(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
A boolean
}
\description{
Test if all values in x are identical
}
\examples{
identical_vals(c(1,1,1,1))
identical_vals(c(1,1,1,2))
}
