% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stics_exe.R
\name{init_javastics_pref}
\alias{init_javastics_pref}
\title{Initialize JavaStics preferences}
\usage{
init_javastics_pref(javastics_path, overwrite = FALSE)
}
\arguments{
\item{javastics_path}{}
}
\value{
\code{TRUE} if the file was created, \code{FALSE} otherwise.
}
\description{
Initialize the JavaStics \code{preferences.xml} file.
}
\examples{
\dontrun{
init_javastics_pref("path/to/javastics")
}

}
\keyword{internal}
