% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_exe_utilities.R
\name{list_stics_exe}
\alias{list_stics_exe}
\title{List stics executables available in JavaStics}
\usage{
list_stics_exe(javastics_path)
}
\arguments{
\item{javastics_path}{Path to the JavaStics installation directory}
}
\value{
A list of two:
\itemize{
\item stics_list: named list of the stics executable
\item current: a named list of the executable currently in use
}
}
\description{
Return all stics identifier names and executable available in JavaStics
}
\details{
The information is read from the \code{preference.xml} file in JavaStics. The function first
controls that it exists and is complete, and if not it creates it using an OS-specific version.
}
\examples{
\dontrun{
 list_stics_exe("path/to/JavaSTICS-v131-stics-v841")
}

}
\keyword{internal}
