% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stics_versions.R
\name{list_stics_versions}
\alias{list_stics_versions}
\title{List stics versions}
\usage{
list_stics_versions(javastics_path)
}
\arguments{
\item{javastics_path}{Path to the JavaStics installation directory}
}
\value{
A list of two:
\itemize{
\item stics_list: named list of the stics executable
\item current: the version currently in use
}
}
\description{
Return all stics versions names and executable listed for use
}
\details{
The standard versins shipped with JavaStics is called modulostics and its
associated executable is named \code{stics_modulo.exe}.
}
\examples{
\dontrun{
list_stics_versions("path/to/JavaSTICS-1.41-stics-9.1")
}
}
