% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_output.R
\name{plot_output}
\alias{plot_output}
\title{Plot STICS outputs}
\usage{
plot_output(..., Vars = NULL)
}
\arguments{
\item{...}{output data.frame(s), generally the output from \code{\link{eval_output}}
to compare with observations. If several objects are detected, make a comparison between
them.}
}
\description{
Plot stics outputs, compare between USMs, and compare simulations
to observations when available.
}
\examples{
\dontrun{
library(sticRs)
plot_output()
}
}
