% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stati_stics.R
\name{predictor_assessment}
\alias{predictor_assessment}
\alias{R2}
\alias{RMSE}
\alias{nRMSE}
\alias{MAE}
\alias{MSE}
\alias{EF}
\alias{Bias}
\alias{ABS}
\alias{NSE}
\alias{FVU}
\alias{RME}
\title{Model quality assessment}
\usage{
R2(sim, obs, na.rm = T)

RMSE(sim, obs, na.rm = T)

nRMSE(sim, obs, na.rm = T)

MAE(sim, obs, na.rm = T)

MSE(sim, obs, na.rm = T)

EF(sim, obs, na.rm = T)

Bias(sim, obs, na.rm = T)

ABS(sim, obs, na.rm = T)

NSE(sim, obs, na.rm = T)

FVU(sim, obs, na.rm = T)

RME(sim, obs, na.rm = T)
}
\arguments{
\item{sim}{Simulated values}

\item{obs}{Observed values}

\item{na.rm}{Boolean. Remove \code{NA} values if \code{TRUE} (default)}
}
\value{
A statistic depending on the function used.
}
\description{
Several statistics to assess the quality of the predictions
             of a model (see note).
}
\details{
The statistics for model quality can differ between sources. Here is a
         list of the methods used in this package for each one (see html
         version for \code{LATEX}):
\describe{
  \item{R2}{Computed using \code{\link[stats]{lm}} on obs~sim}
  \item{RMSE}{\eqn{\sqrt{\frac{\sum_1^n(\hat{y_i}-y_i)^2}{n}}}{
  sqrt(mean((sim-obs)^2)}}
  \item{nRMSE}{\eqn{\frac{RMSE}{\hat{y}}\cdot100}{(RMSE/mean(obs))*100}}
  \item{MAE}{\eqn{\frac{\sum_{i=1}^n\left| y_i-x_i\right|}{n}}{
  mean(abs(sim-obs))}}
  \item{FVU}{\eqn{\frac{SS_(res)}{SS_(tot)}}{SS_res/SS_tot} see
  \href{https://en.wikipedia.org/wiki/Fraction_of_variance_unexplained}{
  wikipedia definition}}
  \item{MSE}{\eqn{\frac{1}{n}\sum_{i=1}^n(Y_i-\hat{Y_i})^2}{
  mean((sim-obs)^2,na.rm = na.rm)}}
  \item{EF}{\eqn{1-\frac{SS_(res)}{SS_(tot)}}{1-SS_res/SS_tot}}
  \item{Bias}{\eqn{\frac{\sum_1^n(\hat{y_i}-y_i)}{n}}{mean(sim-obs)}}
  \item{ABS}{\eqn{\frac{\sum_1^n(\left|\hat{y_i}-y_i\right|)}{n}}{
  mean(abs(sim-obs))}}
  \item{RME}{\eqn{\frac{\sum_1^n(\frac{\hat{y_i}-y_i}{y_i})}{n}}{mean(sim-obs)}}
}
}
\note{
The model efficiency (EF or NSE) shares the same equation as the standard R2,
      except SStot is applied relative to the identity function (1:1 line) instead
      of the regression line. Furthermore, the R2 here is the one from the regression
      applied to the sim~obs relationship, not a direct R2.
}
\examples{
library(sticRs)
sim= rnorm(n = 5,mean = 1,sd = 1)
obs= rnorm(n = 5,mean = 1,sd = 1)
RMSE(sim,obs)
}
\seealso{
This function is largely inspired from the \code{evaluate()} function
         from the \code{SticsEvalR} package
}
