% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_eval.R
\name{rbind_sim}
\alias{rbind_sim}
\title{Makes one stics_eval simulation list from several}
\usage{
rbind_sim(...)
}
\arguments{
\item{...}{Input simulations lists from \code{\link[=stics_eval]{stics_eval()}}.}
}
\description{
Uses \code{\link[data.table:rbindlist]{data.table::rbindlist()}} but keep the original
structure of the list within the list. To use for row binding
\code{\link[=stics_eval]{stics_eval()}} outputs.
}
\details{
The function keep all columns from the simulations as in \code{data.table::rbindlist(l,fill= TRUE)}.
}
\examples{
\dontrun{

Simulation= rbind_sim(Sim_2017, Sim_2018)

}
}
\seealso{
\code{\link[=stics_eval]{stics_eval()}}
}
