% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_met.R
\name{read_met}
\alias{read_met}
\title{Read STICS input meteo file}
\usage{
read_met(dirpath = getwd(), filename = "climat.txt")
}
\arguments{
\item{dirpath}{Directory path}

\item{filename}{The meteorology file name (default to \code{climat.txt}).}
}
\value{
A data.frame of the input meteorological variables used as input for the
        STICS model.
}
\description{
Read the meteorology input for STICS ("climat.txt")
}
\note{
The time-related variables are summarised into one POSIXct column named
      code{Date}.
}
\examples{
\dontrun{
library(sticRs)
Meteo= read_met()
}

}
\seealso{
\code{\link{read_obs}}
}
