% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_obs.R
\name{read_obs}
\alias{read_obs}
\title{Read STICS observation file (*.obs)}
\usage{
read_obs(dirpath = getwd(), filename = NULL, mixed = NULL)
}
\arguments{
\item{dirpath}{Directory path}

\item{filename}{A vector of observation file name(s). Optional, see details.}

\item{mixed}{(optional) Is the simulation on mixed species (boolean)}
}
\value{
A data.frame (sole crop) or a list of two data.frames (mixed crops) of
        the STICS-formated observations. Return \code{NULL} if no files were found,
        or more files than useable.
}
\description{
Read STICS observation file for sole or mixed crops.
}
\details{
If \code{filename} is not specified (or equal to \code{NULL}), the
         function tries first to match the \code{mod_s*} names for the same \*.obs
         names, and then to use the \code{.obs} file if there is one only
         (used for sole crops or for both mixed crops). If there are no .obs files,
         or two but not specified for reading, the function returns \code{NULL}
         If \code{mixed} is not specified (or equal to \code{NULL}), the function
         tries to read the number of species from the input files.
}
\examples{
\dontrun{
library(sticRs)
Meas= read_obs()
}

}
\seealso{
\code{\link{read_output}}
}
