% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{read_output}
\alias{read_output}
\title{Read STICS outputs (mod_s*)}
\usage{
read_output(dirpath = getwd(), mixed = NULL)
}
\arguments{
\item{dirpath}{Directory path}

\item{mixed}{(optional) Is the simulation on mixed species (boolean)}
}
\value{
A data.frame (sole crop) or a list of two data.frames (mixed crops) of
        the STICS outputs.
}
\description{
Read STICS model outputs for sole or mixed crops.
}
\details{
If \code{mixed} is not specified (or equal to \code{NULL}), the function try to
         read the number of species from the input files.
}
\note{
The STICS outputs are controled from the \code{var.mod} file.
      \code{\link{set_out_var}} can be used to set the output variables.
}
\examples{
\dontrun{
library(sticRs)
Output= read_output()
}

}
\seealso{
\code{\link{read_param}}, \code{\link{set_param}}.
}
