% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_param.R
\name{read_param}
\alias{read_param}
\title{Read STICS input parameters}
\usage{
read_param(filepath, param = NULL)
}
\arguments{
\item{filepath}{File path}

\item{param}{Parameter name. Optional, if not provided, the function
return an object with all parameters}
}
\description{
Read an input parameter from a pre-existing STICS input
             file. Generally used after calling \code{\link{set_usm}}.
}
\examples{
\dontrun{
# Replace the interrow distance parameter to 0.01:

library(sticRs)
filepath= '1-Simulations/IC_Wheat_Wheat/fictec1.txt'
Param= read_param(filepath = filepath)

}

}
\seealso{
\code{\link{set_usm}}.
}
