% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_param.R
\name{read_tec}
\alias{read_tec}
\title{Read STICS technical parameters}
\usage{
read_tec(filepath, several_fert = T, several_thin = T, is_pasture = F)
}
\arguments{
\item{filepath}{File path}

\item{several_fert}{Is there several fertilization in the USM ?}

\item{several_thin}{Is there several thinning in the USM ?}

\item{is_pasture}{Is the plant a pasture ?}
}
\description{
Read the technical input parameter from a pre-existing STICS input
             file. Generally not used, prefer using \code{\link{read_param}}.
}
\examples{
\dontrun{
# Replace the interrow distance parameter to 0.01:

library(sticRs)
filepath= '1-Simulations/IC_Wheat_Wheat/fictec1.txt'
Param= read_param(filepath = filepath)

}

}
\seealso{
\code{\link{read_param}}.
}
