% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_java_model.R
\name{remove_java_model}
\alias{remove_java_model}
\title{Removing a Stics model version to use from JavaStics}
\usage{
remove_java_model(javastics_path, java_model_tag)
}
\arguments{
\item{javastics_path}{JavaStics installation root folder}

\item{java_model_tag}{Model name (not executable file name)}
}
\description{
Removing a model entry in preferences file (model exe name, model tag name)
}
\details{
The executable file in the current JavaStics bin folder will not be deleted

It will not be possible to run it through JavaStics command line interface
}
\examples{
\dontrun{
remove_java_model("/home/plecharpent/Work/JavaSTICS-v131-stics-v841","model_name")
}
}
