% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_exe_utilities.R
\name{remove_stics_exe}
\alias{remove_stics_exe}
\title{Remove a Stics model executable from JavaStics}
\usage{
remove_stics_exe(javastics_path, stics_exe)
}
\arguments{
\item{javastics_path}{Path to the JavaStics installation directory}

\item{stics_exe}{Stics executable identifier name (see details)}
}
\value{
Nothing. Update the "preference.xml" file in the config of JavaStics.
}
\description{
Remove a stics model from the list of available model executables in
JavaStics (modifies the "preferences.xml" file).
}
\details{
The executable file in the current JavaStics bin folder will not be deleted,
please do it by hand instead (in "JavaStics.../bin").
The \code{stics_exe} is \strong{not} the name of the executable file, but the
identification name. Please use \code{list_stics_exe()} to list all available executables,
and \code{set_stics_exe()} to add a new one. The identification names can be retreived using
\code{names(list_stics_exe(javastics_path)$stics_list)}
}
\examples{
\dontrun{
remove_stics_exe("path/to/JavaSTICS-v131-stics-v841","model_name")
}

}
\keyword{internal}
