% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_javastics.R
\name{run_javastics}
\alias{run_javastics}
\title{Running one or several usms from a javastics workspace}
\usage{
run_javastics(
  javastics,
  workspace = NULL,
  usm = NULL,
  keep_history = TRUE,
  optim = FALSE,
  verbose = TRUE,
  stics_exe = "modulostics",
  java_cmd = "java",
  javastics_path = lifecycle::deprecated(),
  workspace_path = lifecycle::deprecated(),
  usms_list = lifecycle::deprecated()
)
}
\arguments{
\item{javastics}{Path of JavaStics}

\item{workspace}{Path of a JavaStics workspace}

\item{usm}{Vector of USM names. Optional, if provided, the function runs
only the given USMs.
If not provided, the function runs all the USMs included in workspace.}

\item{keep_history}{Logical value (optional) to keep a copy of history file
use \code{TRUE} (default), \code{FALSE} otherwise}

\item{optim}{Logical value (optional), \code{TRUE} to force code_optim value to 1,
\code{FALSE} otherwise (default)}

\item{verbose}{Logical value for displaying information while running}

\item{stics_exe}{The name, executable or path of the stics executable to use
(optional, default to "modulostics", see details)}

\item{java_cmd}{The java virtual machine command name or executable path}

\item{javastics_path}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{javastics_path}
is no longer supported, use \code{javastics} instead.}

\item{workspace_path}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{workspace_path}
is no longer supported, use \code{workspace} instead.}

\item{usms_list}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{usms_list} is no
longer supported, use \code{usm} instead.}
}
\value{
A list in which each element contains: usm "name", "error" status
(logical) and an output "message" (JavaStics commandline execution output)
}
\description{
This function uses basically Stics through his
JavaStics command line interface
}
\details{
\code{stics_exe} may be :
\enumerate{
\item a model name pointing to a stics executable as done in JavaStics, e.g.
"modulostics" for \code{stics_modulo.exe}, the standard version of the model
shipping with JavaStics;
\item a stics executable file available from the bin folder in JavaStics, e.g.
"stics_modulo.exe";
\item a path to a stics executable file, eg.
"C:/Users/username/Desktop/stics.exe". NB: this file cannot be named
\code{stics_modulo.exe} because it is the name of the standard STICS shipping
with JavaStics (overwriting is not allowed).
}
}
\examples{
\dontrun{
run_javastics("/path/to/JavaSTICS/folder", "example")
run_javastics("/path/to/JavaSTICS/folder", "/path/to/workspace")
run_javastics("/path/to/JavaSTICS/folder", "example", c("wheat", "pea"))
run_javastics("/path/to/JavaSTICS/folder", usm = c("wheat", "pea"))
run_javastics("/path/to/JavaSTICS/folder",
  usm = c("wheat", "pea"), optim = TRUE
)
}

}
