% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_system.R
\name{run_system}
\alias{run_system}
\title{Running usm(s) from a/several directory(ies)
or a/several subdirectory(ies) named with usm name}
\usage{
run_system(model_path, data_dir, usm_dir_names = NULL)
}
\arguments{
\item{model_path}{Path of Stics executable file}

\item{data_dir}{Path of a Stics input directory or a vector of,
or root directory of Stics directories (if usm_dir_names is given)}

\item{usm_dir_names}{Name(s) of sub-directory(ies) of data_dir}
}
\value{
A list with usm names and execution error status
}
\description{
This function uses Stics directly through a system call
}
\examples{
run_system("/home/username/bin/Stics","/home/username/Work/SticsInputsDir")
run_system("/home/username/bin/Stics",c("/home/username/Work/SticsInputsDir1",
"/home/username/Work/SticsInputsDir2"))

run_system("/home/username/bin/Stics",/home/username/Work/SticsInputsRootDir","wheat")
run_system("/home/username/bin/Stics",/home/username/Work/SticsInputsRootDir",
c("wheat","maize"))

}
