% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_system_cmd.R
\name{run_system_cmd}
\alias{run_system_cmd}
\title{Run a system command}
\usage{
run_system_cmd(command, args = "", output = FALSE)
}
\arguments{
\item{command}{A system command}

\item{args}{A list of arguments to pass to the command}

\item{output}{A logical value to specify returning (T) or not (F, default)
the command output in the function return "output" attribute}
}
\value{
A logical status T if it succeded or F if not
a "message" attribute is set with an error message, and optionaly
an "output" attribute is set with the command output
}
\description{
Run a system command
}
\keyword{internal}
