% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_read_sensitivity.R
\name{save_sensi}
\alias{save_sensi}
\alias{load_sensi}
\alias{load_sensi}
\title{Save and Read sensitivity analysis of STICS}
\usage{
save_sensi(x, dirpath = getwd(), device = "png", width, height, units, ...)

load_sensi(dirpath = getwd())
}
\arguments{
\item{x}{The output from the \code{\link{sensitive_stics}} function}

\item{dirpath}{Path to the directory where to save the plots and objects}

\item{device}{Device to use for the \code{\link[ggplot2]{ggsave}} function
that is used under the hood. From its documentation: Can be either
a device function (e.g. png()), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only)}

\item{width, height}{Plot dimensions in \code{units} parameter units}

\item{units}{Units to use for the plot size parameters
them (see details and example)}

\item{...}{Further parameters to pass to \code{\link[ggplot2]{ggsave}}}
}
\description{
Helper functions to conveniently read or save sensitivity outputs from
             the \code{\link{sensitive_stics}} function
}
\details{
The function first save all the `gg_objects` from the output of
         \code{\link{sensitive_stics}} as images, save the full `x`object using
         \code{\link[base]{saveRDS}}, and write a readme file for convenience.
}
\examples{
\dontrun{
library(sticRs)

# Making a sensitivity analysis on two parameters:

sens=
  sensitive_stics(dir.orig = "0-DATA/dummy/SC_Wheat",
                  dir.targ = "Sensitivity",
                  stics = "0-DATA/stics_executable/original/stics.exe",
                  obs_name = "Wheat_N0.obs",
                  Parameters= list(interrang= list(min=0.05, max=0.25),
                                   P_densitesem= list(min=140, max=280)),
                  Vars = c("raint", "lai(n)", "masec(n)"),
                  method= "fast99", n= 10,
                  q= "qunif",q.arg = list(list(min=0.05, max=0.25),
                  list(min=140, max=280)))

# Saving the output :

save_sensi(x = sens,dirpath = "3-Results/Sensitivity",
           width = 16,height = 9,units = "in")

sens= load_sensi(dirpath = "3-Results/Sensitivity")
}
}
\seealso{
\code{\link[ggplot2]{ggsave}}, \code{\link[base]{saveRDS}} and
         \code{\link[base]{readRDS}}
}
