% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stics_exe.R
\name{select_stics_exe}
\alias{select_stics_exe}
\title{Select the Stics executable}
\usage{
select_stics_exe(javastics_path, stics_exe = "stics_modulo")
}
\arguments{
\item{javastics_path}{Path to the JavaStics installation directory}

\item{stics_exe}{Stics executable name (see details)}
}
\value{
Nothing. Update the "preference.xml" file in the config of JavaStics.
}
\description{
Select the Stics model executable to use from the preference file in JavaStics.
}
\details{
The \code{stics_exe} is \strong{not} the name of the executable file, but the
identification name. Please use \code{list_stics_exe()} to list all available executables,
and \code{set_stics_exe()} to add and select a new one. The identification names can be retreived using
\code{names(list_stics_exe(javastics_path)$stics_list)}
}
\note{
"stics_modulo", "sticsmodulo" and "modulostics" are synonyms for the standard STICS executable.
}
\examples{
\dontrun{
select_stics_exe("path/to/JavaSTICS-v131-stics-v841", "modulostics")
}

}
