% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitive_stics.R
\name{sensitive_stics}
\alias{sensitive_stics}
\title{Make sensitivity analysis of STICS output}
\usage{
sensitive_stics(dir.orig, dir.targ = getwd(), stics, obs_name, Parameters,
  Vars, method = c("fast99", "sobol"), Plant = 1, ...)
}
\arguments{
\item{dir.orig}{Path to the directory from which to copy the simulation files. If
\code{NULL} (the default), uses the package dummy USM.}

\item{dir.targ}{Path to the target directory for evaluation. Created if missing.}

\item{stics}{STICS executable path}

\item{obs_name}{A vector of observation file name(s). It must have the form
\code{c(Dominant,Dominated)} for mixed crops.
See \code{\link{read_obs}} \code{filename} parameter for more details.}

\item{Parameters}{A vector of input parameter names}

\item{Vars}{Output variables on which the sensitivity is performed}

\item{method}{The sensitivity method to use, see \pkg{sensitivity} package}

\item{Plant}{On which plant (\emph{i.e.} Principal or associated) the parameters
have to be set (only for mixed simulations, using plant parameters)}

\item{...}{Further parameters to give to the sensitivity function used
(see \pkg{sensitivity} package)}
}
\value{
The sensitivity analysis output
}
\description{
Make sensitivity analysis on a particular output given STICS input
}
\examples{
\dontrun{
library(sticRs)
# Example using the fast99 method:
sensitive_stics(dir.orig = "0-DATA/dummy/SC_Wheat",
                dir.targ = "2-Simulations/Sensitivity2",
                stics = "0-DATA/stics_executable/EquDens_trg/stics.exe",
                obs_name = "Wheat_N0.obs",Parameters = "interrang",
                Vars = c("raint", "lai(n)", "masec(n)"),
                method= "fast99", n= 10,
                q= "qunif",q.arg = list(list(min=0.05, max=0.25),
                list(min=140, max=280)))
}

}
\seealso{
\code{\link[stats]{Distributions}} if you use \code{\link[sensitivity]{fast99}}.
}
