% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_usm.R
\name{set_file}
\alias{set_file}
\title{Internal function to set some STICS input file parameters}
\usage{
set_file(filepath, param, value, add)
}
\arguments{
\item{filepath}{Path to the parameter file}

\item{param}{Parameter name}

\item{value}{New parameter value}

\item{add}{Boolean. Append input to existing file (add to the list)}
}
\description{
Replace or set an input parameter from a pre-existing STICS input
             file. This function is called by some of the generic \code{set_*}
             functions under the hood.
}
\details{
The function uses \code{\link[base]{sys.call}} to know from which function
         of the \code{set_*} family it is called, so it won't work properly if called
         by the user directly. This is why this function is internal.
}
\note{
This function is not used for \code{\link{set_soil}}.
}
\seealso{
\code{\link{set_param}}.
}
\keyword{internal}
