% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_file_executable.R
\name{set_file_executable}
\alias{set_file_executable}
\title{Setting a file to executable access mode}
\usage{
set_file_executable(file_path)
}
\arguments{
\item{file_path}{A file path}
}
\value{
Execution status: TRUE for successfull operation, FALSE otherwise
}
\description{
Setting executable status on for a file under Linux or Mac OS.
}
\keyword{internal}
