% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_java_model.R
\name{set_java_model}
\alias{set_java_model}
\title{Selecting Stics model to use}
\usage{
set_java_model(javastics_path, java_model_tag)
}
\arguments{
\item{javastics_path}{JavaStics installation root folder}

\item{java_model_tag}{Model name (not executable file)}
}
\description{
Setting in preferences the Stics model version to use
}
\examples{
set_java_model("/home/plecharpent/Work/JavaSTICS-v131-stics-v841","model_name")

}
