% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_java_pref.R
\name{set_java_pref}
\alias{set_java_pref}
\title{Setting default preferences}
\usage{
set_java_pref(javastics_path)
}
\arguments{
\item{javastics_path}{JavaStics installation root folder}
}
\value{
Current Stics model name used in JavaStics default configuration
}
\description{
According to system detection generating a default preference file for a OS specific executable
}
\examples{
\dontrun{
current_model <- set_java_pref("/home/plecharpent/Work/JavaSTICS-v131-stics-v841")
}

}
