% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_java_workspace.R
\name{set_java_workspace}
\alias{set_java_workspace}
\title{Setting JavaStics workspace}
\usage{
set_java_workspace(javastics_path, java_wd)
}
\arguments{
\item{javastics_path}{JavaStics installation root folder}

\item{java_wd}{JavaStics working directory (absolute,relative to javastics_path)}
}
\description{
Setting a new JavaStics working directory, a relative directory to JavaStics path or an absolute one
}
\details{
Checking if the directory is a JavaStics workspace (any usms.xml file), and if it's already registered

before setting new one
}
\examples{
\dontrun{
 set_java_wd("path/to/JavaSTICS","my_wd")
 set_java_wd("path/to/JavaSTICS","/path/to/my_wd")
}

}
\keyword{internal}
