% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_wrapper.R
\name{set_out_var_internal}
\alias{set_out_var_internal}
\title{Set the output variables needed}
\usage{
set_out_var_internal(
  filepath = "var.mod",
  vars = c("lai(n)", "masec(n)"),
  add = F
)
}
\arguments{
\item{filepath}{The path to the "var.mod" file.}

\item{vars}{The variables needed}

\item{add}{Do we want to add or append the variables to existing variables ?}
}
\value{
Nothing, write in the "var.mod" file
}
\description{
Set the output variables needed
}
\note{
This is exactly the same function as \verb{[sticsRFiles::set_out_var_txt()]}, but we add it here
to avoid a dependency just for a one-liner function.
}
\keyword{internal}
