% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_usm.R
\name{set_param}
\alias{set_param}
\title{Replace STICS input file parameter}
\usage{
set_param(filepath, param = "interrang", value = 0.01)
}
\arguments{
\item{filepath}{File path}

\item{param}{Parameter name}

\item{value}{New value}
}
\description{
Replace or set an input parameter from a pre-existing STICS input
             file. Generally used after calling \code{\link{set_usm}}.
}
\examples{
\dontrun{
# Replace the interrow distance parameter to 0.01:

library(sticRs)
set_param(filepath = "fictec1.txt", param= "interrang", value= 0.01)

}

}
\seealso{
\code{\link{set_usm}}.
}
