% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_stics_exe.R
\name{set_stics_exe}
\alias{set_stics_exe}
\title{Set the Stics executable}
\usage{
set_stics_exe(javastics_path, stics_exe = "stics_modulo")
}
\arguments{
\item{javastics_path}{Path to the JavaStics installation directory}

\item{stics_exe}{Stics executable name (see details)}
}
\value{
Nothing. Update the "preference.xml" file in the config of JavaStics.
}
\description{
Set the Stics model executable to use in JavaStics.
}
\details{
The \code{stics_exe} is \strong{not} the name of the executable file, but the
identification name. Please use \code{list_stics_exe()} to list all available executables,
and \code{add_stics_exe()} to add a new one. The identification names can be retreived using
\code{names(list_stics_exe(javastics_path)$stics_list)}
}
\examples{
\dontrun{
set_stics_exe("path/to/JavaSTICS-v131-stics-v841", "modulostics")
}

}
