% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_stics_version.R
\name{set_stics_version}
\alias{set_stics_version}
\title{Change the version of stics}
\usage{
set_stics_version(javastics_path, stics = "stics_modulo")
}
\arguments{
\item{javastics_path}{Path to the JavaStics installation directory}

\item{stics}{The name of the stics version to use (as given in the
"Stics model selection" tab) from JavaStics}
}
\value{
Nothing. Modifies the \code{preferences.xml} in the config.
}
\description{
Set the version of stics used by JavaStics. The versions available
can be listed using \code{list_stics_versions()}.
}
\note{
To add a new version, put the executable in the bin directory of JavaStics, then
open JavaStics, press \code{ctrl+M}, go to the "Add" tab and add the new version.
}
\examples{
\dontrun{
 set_stics_version("path/to/JavaSTICS-1.41-stics-9.1", "stics_intercrop")
}
}
