% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_usm.R
\name{set_usm}
\alias{set_usm}
\title{Create a simulation unit (USM) for the STICS model}
\usage{
set_usm(dir.orig = NULL, dir.targ = getwd(), dir.stics = NULL,
  usm_name = NULL)
}
\arguments{
\item{dir.orig}{Path to the directory from which copy the simulation files. If
\code{NULL} (the default), uses the package dummy USM.}

\item{dir.targ}{Path to the target directory for simulation. Created if missing.}

\item{dir.stics}{Path to the STICS model executable (optional, only needed if not present
in dir.orig)}

\item{usm_name}{Vector name of the USM(s).}
}
\description{
Uses a pre-existing USM, copy files to target folder(s) of simulation,
             set the USMs ready for simulation. \code{set_usm} create as many sub-folders
             as USM neede in the target folder.
}
\details{
This function is a helper function used by other package functions
}
\examples{
\dontrun{
# Set a dummy simulation in the folder "1-Simulations", relative to the
# project path:

library(sticRs)
set_usm(dir.targ = "1-Simulations")

}

}
