% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stati_stics.R
\name{stati_stics}
\alias{stati_stics}
\title{Compute simulated/observed statistics}
\usage{
stati_stics(..., obs_name = NUL)
}
\arguments{
\item{...}{Either a file path or the output from \code{\link{eval_output}}. If several objects are
detected, make a comparison between them.}

\item{obs_name}{A vector of observation file name(s). Optionnal, required only if \code{...} is a link.
Must have the form \code{c(Dominant,Dominated)} for mixed crops. See details.}
}
\value{
A data.frame with statistics for each simulation.
}
\description{
Compute statistics for evaluation of the STICS outputs against observations. This function can
             be used for one USM or to compare outputs from different model versions or parameter values.
}
\details{
If \code{obs_name} is not provided, the function tries to guess it using the built-in algorithm
         from \code{\link{read_obs}}. See respective documentation for more details.
}
\examples{
\dontrun{
library(sticRs)

}
}
\seealso{
This function is largely inspired from the \code{evaluate()} function from the
         \code{SticsEvalR} package
}
