% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_eval.R
\name{stics_eval}
\alias{stics_eval}
\title{Evaluate STICS version}
\usage{
stics_eval(dir.orig = NULL, dir.targ = getwd(), stics, obs_name = NULL,
  Out_var = NULL, plot_it = T, Parallel = T, mixed = NULL)
}
\arguments{
\item{dir.orig}{Path to the directory from which to copy the simulation files. If
\code{NULL} (the default), uses the package dummy USM.}

\item{dir.targ}{Path to the target directory for evaluation. Created if missing.}

\item{stics}{STICS executable path named list.}

\item{obs_name}{A vector of observation file name(s). It must have the form
\code{c(Dominant,Dominated)} for mixed crops.
See \code{\link{read_obs}} \code{filename} parameter for more details.}

\item{Out_var}{The variables needed as output}

\item{plot_it}{Boolean. Do the plot as to be pinted ?}

\item{Parallel}{Are the simulations to be executed in parallel ?}

\item{mixed}{(optional) Is the simulation made on mixed species (boolean)}
}
\value{
A list of two. The first object is a list of the outputs of
 \code{\link{eval_output}}, which is called for each object in \code{stics}.
 The second is a summary of the simulations outputs as a ggplot object.
}
\description{
Wrapper function to run STICS simulations and evaluate
 outputs using observations and/or other model versions.
}
\details{
The names in the \code{stics} parameter list are used for reference in the outputs
(data.frames and plots). If no names are provided, the function give a dummy name for each.
The function run STICS executable in parrallel using all cores from the
machine but one.
}
\examples{
\dontrun{
library(sticRs)
stics_eval()
}

}
\seealso{
\code{\link{eval_output}}, \code{\link{import_usm}},
 \code{\link{run_stics}}, and \code{\link{set_out_var}}
}
