% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimi_stics.R
\name{stics_eval_opti}
\alias{stics_eval_opti}
\title{Objective function to optimize}
\usage{
stics_eval_opti(x, USM_path, obs_name, param, weight = NULL, Plant)
}
\arguments{
\item{x}{The starting parameter values}

\item{USM_path}{The path to the USM}

\item{obs_name}{The observation file name}

\item{param}{The parameter names (in same order than x)}

\item{weight}{The weight used for each variable (see details)}

\item{Plant}{The plant (\emph{i.e.} Principal or associated) for which the parameters
will be set (only for plant or technical parameters in mixed crop simulations)
Set to \code{NULL} if using STICS in sole crop}
}
\value{
The weighted product of squares (selection criteria)
}
\description{
Function evaluated for parameter optimization. This function is only provided
for informative value, but should not be used by the user. It is called by \code{\link[=optimi_stics]{optimi_stics()}}.
}
\details{
If weight is not provided by the user, the selection criteria is computed using the equation
5 from Wallach et al. (2011). If they are provided, the equation 6 is used instead.
}
\references{
Wallach, D., Buis, S., Lecharpentier, P., Bourges, J., Clastre, P., Launay, M., … Justes, E. (2011).
A package of parameter e00stimation methods and implementation for the STICS crop-soil model. Environmental Modelling & Software, 26(4), 386–394. doi:10.1016/j.envsoft.2010.09.004
}
