% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stics.R
\name{check_stics}
\alias{check_stics}
\title{Checking if given path is a Stics executable path}
\usage{
check_stics(model_path, version = FALSE, stop = TRUE)
}
\arguments{
\item{model_path}{Model executable path}

\item{version}{Logical, or getting system command return i.e.
model version or not (default)}
}
\value{
System output (error,...)
}
\description{
Checking if given path is a Stics executable path
}
