% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimi_stics.R
\name{optimi_stics}
\alias{optimi_stics}
\title{STICS parameter optimization}
\usage{
optimi_stics(dir.orig, dir.targ = getwd(), stics, obs_name, Parameters,
  Vars, weight = NULL, method = c("nmkb"), Plant = 1, ...)
}
\arguments{
\item{dir.orig}{Vector or named list of paths to the directory from which to copy the USMs files.}

\item{dir.targ}{Path to the target directory for evaluation. Created if missing.}

\item{stics}{STICS executable path}

\item{obs_name}{A `data.frame` of observation file name(s) of the form
`data.frame(Principal= c(obs1.obs), Dominated= c(obs2.obs)` for mixed crops (simply remove the `Dominated` column for sole crops.}

\item{Parameters}{A data.frame with parameter name, starting, min and max values (see details and example)}

\item{Vars}{Output variables on which the optimization is performed}

\item{weight}{The weight used for each variable (see details)}

\item{method}{The optimization method to use, see \pkg{dfoptim} package. For the moment, only [dfoptim::nmkb()]}

\item{Plant}{A vector for the plant (*i.e.* Principal or associated) for which the parameters
will be set (only for plant or technical parameters in mixed crop simulations)
Set to `NULL` if using STICS in sole crop}

\item{...}{Further parameters passed to the optimization function called
(see \pkg{dfoptim} package)}
}
\value{
A list of three :
* gg_objects: A list of ggplot objects to plot the final STICS simulation with optimized parameter values
  compared to original parameter values.
* opti_output: differs if the optimisation is uni or multi-dimensional:
    + Univariate: A list of two: `minimum` -> the optimized parameter(s) value(s), and
      `objective` -> the last value of the objective function.
    + Multivariate: A list of six: `par` -> the optimized parameter(s) value(s),
      `value` -> the last value of the optimization function, `feval`-> the number of times the objective function
       was evaluated, `restarts` -> the number of times the algorithm had to be restarted when it stagnated,
       `convergence` -> 0= convergence, 1 no convergence, and `message` some information about the convergence.
* last_sim_data: the STICS output from the last simulation with the optimized parameter
  values. This data is given using [eval_output()].
}
\description{
Optimize STICS parameter values according to measurements
}
\details{
The function uses [stats::optimize()] for univariate optimization, and the \pkg{dfoptim} package functions for multivariate.
Currently only the Nelder-Mead algorithm is implemented from \pkg{dfoptim}.
The `Parameters` argument should be formated as a a data.frame (see example).
The start values should exclude the min and max values (they are exclusive bounds).
If the start is `NULL`, then the mean value between the min and max values is taken.
If weight is not provided by the user, the selection criteria is computed using the equation
5 from Wallach et al. (2011). If they are provided, the equation 6 is used instead.
}
\examples{
\dontrun{
library(sticRs)

Parameters= data.frame(parameter= c('hautK1','hautK2'),
                       start= c(0.2,0.2),
                       min= c(0,0),
                       max= c(1,1))

# On a single USM:
optimi_stics(dir.orig = "0-DATA/dummy/Year_2005_2006/IC_Wheat_Pea",
             dir.targ = "2-Simulations/param_optim",
             stics = "0-DATA/stics_executable/19-new/Stics.exe",
             obs_name = c("6_IC_Wheat_N0.obs","6_IC_Pea_N0.obs"),
             Parameters= Parameters,
             Vars = c('lai(n)','masec(n)','hauteur'),
             method= "nmkb",Plant=1)

# On a series of USMs:
optimi_stics(dir.orig = list(Y2005= "0-DATA/dummy/Year_2005_2006/IC_Wheat_Pea",
                             Y2006= "0-DATA/dummy/Year_2006_2007/IC_Wheat_Pea"),
             dir.targ = "2-Simulations/param_optim",
             stics = "0-DATA/stics_executable/19-new/Stics.exe",
             obs_name = data.frame(Principal= rep("6_IC_Wheat_N0.obs",2),
                                   Associated= rep("6_IC_Pea_N0.obs",2)),
             Parameters = Parameters, weight= 1, Vars = c('hauteur'),
             method= "nmkb",Plant=c(1,1))
}

}
\references{
Wallach, D., Buis, S., Lecharpentier, P., Bourges, J., Clastre, P., Launay, M., … Justes, E. (2011).
 A package of parameter estimation methods and implementation for the STICS crop-soil model. Environmental Modelling & Software, 26(4), 386–394. doi:10.1016/j.envsoft.2010.09.004
}
\seealso{
[dfoptim::nmkb()]
}
