% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_output.R
\name{plot_output}
\alias{plot_output}
\title{Plot STICS outputs}
\usage{
plot_output(..., Vars = NULL, obs_name = NULL)
}
\arguments{
\item{...}{Either a file path or the output from \code{\link{eval_output}}
If several objects are detected, make a comparison between them.}

\item{Vars}{Character vector of variables required for plotting.}

\item{obs_name}{A vector of observation file name(s). It must have the form
\code{c(Dominant,Dominated)} for mixed crops. See details.}
}
\description{
Plot stics outputs, compare between USMs, and compare simulations
to observations when available.
}
\details{
if \code{Vars} is NULL (the default), the function plots all variables
from the simulation. The output variables from simulations can be set using
\code{\link{set_out_var}}.If \code{obs_name} is not provided, the function try
to guess it using the built-in algorithm from \code{\link{read_obs}}. Idem for
the \code{mixed} argument. See documentation for more details.
}
\examples{
\dontrun{
library(sticRs)
plot_output()
}
}
