% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_output.R
\name{plot_output}
\alias{plot_output}
\title{Plot STICS outputs}
\usage{
plot_output(
  ...,
  Vars = NULL,
  obs_name = NULL,
  Title = NULL,
  plot_it = T,
  group = F
)
}
\arguments{
\item{...}{Either a file path or the output from \code{\link[=eval_output]{eval_output()}}
If several objects are detected, make a comparison between them.}

\item{Vars}{Character vector of variables required for plotting.}

\item{obs_name}{A vector of observation file name(s). It must have the form
\code{c(Dominant,Dominated)} for mixed crops. See details.}

\item{Title}{A title for the plot (optional)}

\item{plot_it}{Boolean. Do the plot as to be pinted ?}

\item{group}{Boolean (Optional). Should the \code{group} column be used for
column faceting. If \code{TRUE}, make sure that the group column
exists in the input}
}
\value{
A ggplot object, and print a plot if \code{plot_it} is set to \code{TRUE}.
}
\description{
Plot stics outputs, compare between USMs, and compare simulations
to observations when available.
}
\details{
if \code{Vars} is NULL (the default), the function plots all variables
from the simulation. The output variables from simulations can be set using
\code{\link[=set_out_var]{set_out_var()}}.If \code{obs_name} is not provided, the function tries
to guess it using the built-in algorithm from \code{\link[=read_obs]{read_obs()}}. See
respective documentation for more details.
}
\examples{
\dontrun{
library(sticRs)
# Example 1, uing paths as inputs:
plot_output(Sim1= "dummy_path/simulation_1",Sim2= "dummy_path/simulation_2",
            obs_name = c("Wheat.obs","Pea.obs"),
            Title = "Model comparison for Wheat-Pea in Intercrop")

# Example 2, using do.call :
Simulations_path= list(Sim1= "dummy_path/simulation_1",Sim2= "dummy_path/simulation_2")
# Add arguments passed to plot_output() in the "Simulations_path" list:
Simulations_path$Title= "Model comparison for Wheat-Pea in Intercrop"
do.call(plot_output,Simulations_path)

# Example 3, using simulation outputs:
sim1= read_output("dummy_path/simulation_1")
sim2= read_output("dummy_path/simulation_2")
plot_output(sim1=sim1,sim2=sim2)

}
}
