% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_param.R
\name{read_param}
\alias{read_param}
\alias{read_ini}
\alias{read_general}
\alias{read_plant}
\alias{read_tec}
\alias{read_soil}
\alias{read_station}
\title{Read STICS input parameters}
\usage{
read_param(dirpath, param = NULL, ...)

read_ini(filepath)

read_general(filepath)

read_plant(filepath, max_variety = 30)

read_tec(filepath, several_fert = T, several_thin = T, is_pasture = F)

read_soil(filepath)

read_station(filepath)
}
\arguments{
\item{dirpath}{Directory path}

\item{param}{Parameter name. Optional, if not provided, the function
return an object with all parameters}

\item{...}{Helper to pass arguments from \code{\link{read_param}} to the
other functions}

\item{filepath}{File path}

\item{max_variety}{Maximum number of variety authorized (this is only for STICS
compatibility)}

\item{several_fert}{Is there several fertilization in the USM ?}

\item{several_thin}{Is there several thinning in the USM ?}

\item{is_pasture}{Is the plant a pasture ?}
}
\value{
A list of all parameters:
        \item{ini}{Initialization parameters}
        \item{general}{General parameters}
        \item{tec}{Technical parameters}
        \item{plant}{Plant parameters}
        \item{soil}{Soil parameters}
        \item{station}{Station parameters}
The function can return several \code{plant} and \code{tec} if mixed crops,
numbered by usage
}
\description{
Read STICS model input parameters from a pre-existing STICS input
             file. Generally used after calling \code{\link{set_usm}}.
}
\note{
Users generally only use \code{read_param} that identify parameters for
      other functions and call them.
}
\examples{
\dontrun{
# Replace the interrow distance parameter to 0.01:

library(sticRs)
filepath= '1-Simulations/IC_Wheat_Wheat/fictec1.txt'
Param= read_param(filepath = filepath)

}

}
\seealso{
\code{\link{set_usm}}.
}
