% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_javastics.R
\name{run_javastics}
\alias{run_javastics}
\title{Running one or several usms from a javastics workspace}
\usage{
run_javastics(
  javastics_path,
  javastics_workspace_path = NULL,
  usms_list = c(),
  keep_history = TRUE,
  optim = FALSE
)
}
\arguments{
\item{javastics_path}{Path of JavaStics installation directory}

\item{javastics_workspace_path}{Path of a JavaStics workspace (Optional)}

\item{usms_list}{Vector of usms to run (Optional)}

\item{keep_history}{Logical value (Optional) to keep a copy of history file
use TRUE (default), FALSE otherwise}

\item{optim}{Logical value (Optional) to force code_optim value to 1, use TRUE,
FALSE otherwise (default)}
}
\value{
A list with usms names and execution error status or NULL ikf  any problem about
the JavaStics workspace or JavaStics installation directory
}
\description{
This function uses basically Stics through his JavasStics command line interface
}
\examples{
\dontrun{
run_java("/home/plecharpent/Work/JavaSTICS-v131-stics-v841","example")
run_java("/home/plecharpent/Work/JavaSTICS-v131-stics-v841","/home/plecharpent/data/example")
run_java("/home/plecharpent/Work/JavaSTICS-v131-stics-v841","example",c("ble","pois"))
run_java("/home/plecharpent/Work/JavaSTICS-v131-stics-v841",usms_list=c("ble","pois"))
run_java("/home/plecharpent/Work/JavaSTICS-v131-stics-v841",usms_list=c("ble","pois"), optim=TRUE)
}

}
