% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_javastics.R
\name{run_javastics}
\alias{run_javastics}
\title{Running one or several usms from a javastics workspace}
\usage{
run_javastics(
  javastics_path,
  workspace_path = NULL,
  usms_list = NULL,
  keep_history = TRUE,
  optim = FALSE,
  display = FALSE,
  stics_exe = "stics_modulo"
)
}
\arguments{
\item{javastics_path}{Path of JavaStics installation directory}

\item{workspace_path}{Path of a JavaStics workspace (Optional)}

\item{usms_list}{Vector of usms to run (optional)}

\item{keep_history}{Logical value (optional) to keep a copy of history file
use TRUE (default), FALSE otherwise}

\item{optim}{Logical value (optional), TRUE to force code_optim value to 1,
FALSE otherwise (default)}

\item{display}{Logical value (optional), TRUE to display usms names,
FALSE otherwise (default)}

\item{stics_exe}{The name of the stics executable to use, default to "stics_modulo" (see details)}
}
\value{
A list in which each element contains: usm "name", "error" status (logical)
and an output "message" (JavaStics commandline execution output)
}
\description{
This function uses basically Stics through his
JavaStics command line interface
}
\details{
Use \code{list_stics_exe()} to list all executables available for the \code{stics_exe} argument.
}
\examples{
\dontrun{
run_javastics("/path/to/JavaSTICS-1.41-stics-9.1","example")
run_javastics("/path/to/JavaSTICS-1.41-stics-9.1","/path/to/workspace")
run_javastics("/path/to/JavaSTICS-1.41-stics-9.1","example",c("wheat","pea"))
run_javastics("/path/to/JavaSTICS-1.41-stics-9.1",usms_list=c("wheat","pea"))
run_javastics("/path/to/JavaSTICS-1.41-stics-9.1",usms_list=c("wheat","pea"), optim=TRUE)
}

}
