% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stics.R
\name{run_stics}
\alias{run_stics}
\title{Run STICS}
\usage{
run_stics(dirpath = getwd())
}
\arguments{
\item{dirpath}{USM directory path}
}
\description{
Helper function to start a simulation using
STICS executable
}
\examples{
\dontrun{
library(sticRs)
run_stics()
}
}
