% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_system.R
\name{run_system}
\alias{run_system}
\title{Running usm(s) from a/several directory(ies)
or a/several subdirectory(ies) named with usm name}
\usage{
run_system(model_path, data_dir, usm_dir_names = NULL, check_exe = TRUE)
}
\arguments{
\item{model_path}{Path of Stics executable file}

\item{data_dir}{Path of a Stics input directory or a vector of,
or root directory of Stics directories (if usm_dir_names is given)}

\item{usm_dir_names}{Name(s) vector of sub-directory(ies) of data_dir
or "all" for extracting all sub-directories path}

\item{check_exe}{Logical, T for checking the model executable, F otherwise}
}
\value{
A list with usm names and execution error status
}
\description{
This function uses Stics directly through a system call
}
\examples{
\dontrun{

# Specifying individual usm directories
run_system("/home/username/bin/Stics","/home/username/Work/SticsInputsDir")
run_system("/home/username/bin/Stics",c("/home/username/Work/SticsInputsDir1",
"/home/username/Work/SticsInputsDir2"))

# Specifying a parent directory of usms directories
# running one or several usms
run_stics("/home/username/bin/Stics","/home/username/Work/SticsInputsRootDir","wheat")
run_stics("/home/username/bin/Stics","/home/username/Work/SticsInputsRootDir",
c("wheat","maize"))
# running all usms
run_stics("/home/username/bin/Stics","/home/username/Work/SticsInputsRootDir",
"all")
}

}
\keyword{internal}
