% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_system_cmd.R
\name{run_system_cmd}
\alias{run_system_cmd}
\title{Run a system command}
\usage{
run_system_cmd(command, args = "", output = FALSE)
}
\arguments{
\item{command}{A system command}

\item{args}{A list of arguments to pass to the command}

\item{output}{A logical value to specify returning the command output (T) or not (F, default)}
}
\value{
A logical status T if it succeded or F if not (an attribute "message" is set with an error message)
}
\description{
Run a system command
}
