% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_javastics_pref.R
\name{set_javastics_pref}
\alias{set_javastics_pref}
\title{Setting default preferences}
\usage{
set_javastics_pref(javastics_path)
}
\arguments{
\item{javastics_path}{Path to the JavaStics installation folder}
}
\value{
Nothing. Creates a default preference file in the config folder of JavaStics
}
\description{
Generates a default preference file ("preferences.xml") according to user's OS
}
\examples{
\dontrun{
current_model <- set_javastics_pref("path/to/JavaSTICS-v131-stics-v841")
}

}
\keyword{internal}
