% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_usm.R
\name{set_param}
\alias{set_param}
\alias{set_usm}
\alias{set_station}
\alias{set_ini}
\alias{set_general}
\alias{set_plant}
\alias{set_tec}
\alias{set_soil}
\title{Replace STICS input file parameter}
\usage{
set_param(dirpath = getwd(), param, value, plant = 1)

set_usm(filepath = "new_travail.usm", param, value)

set_station(filepath = "station.txt", param, value)

set_ini(filepath = "ficini.txt", param, value)

set_general(filepath = "tempopar.sti", param, value)

set_plant(filepath = "ficplt1.txt", param, value)

set_tec(filepath = "fictec1.txt", param, value)

set_soil(filepath = "param.sol", param, value)
}
\arguments{
\item{dirpath}{USM directory path}

\item{param}{Parameter name}

\item{value}{New parameter value}

\item{plant}{Plant index. Optional, only for plant or technical parameters}

\item{filepath}{Path to the parameter file}
}
\description{
Replace or set an input parameter from a pre-existing STICS input
             file. Generally used after calling \code{\link{import_usm}}.
}
\examples{
\dontrun{
# Replace the interrow distance parameter to 0.01:

library(sticRs)
set_param(dirpath = "stics_usm/usm_1", param= "interrang", value= 0.01)

}

}
\seealso{
\code{\link{import_usm}}.
}
