% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_usm.R
\name{set_param}
\alias{set_param}
\alias{set_usm}
\alias{set_station}
\alias{set_ini}
\alias{set_general}
\alias{set_tmp}
\alias{set_plant}
\alias{set_tec}
\alias{set_soil}
\alias{set_out_var}
\title{Set (replace) STICS input file parameters}
\usage{
set_param(dirpath = getwd(), param, value, add = F, plant = 1)

set_usm(filepath = "new_travail.usm", param, value, add = F)

set_station(filepath = "station.txt", param, value, add = F)

set_ini(filepath = "ficini.txt", param, value, add = F)

set_general(filepath = "tempopar.sti", param, value, add = F)

set_tmp(filepath = "tempoparv6.sti", param, value, add = F)

set_plant(filepath = "ficplt1.txt", param, value, add = F, variety = NULL)

set_tec(filepath = "fictec1.txt", param, value, add = F)

set_soil(filepath = "param.sol", param, value)

set_out_var(filepath = "var.mod", vars = c("lai(n)", "masec(n)"), add = F)
}
\arguments{
\item{dirpath}{USM directory path}

\item{param}{Parameter name}

\item{value}{New parameter value}

\item{add}{Boolean. Append input to existing file (add to the list)}

\item{plant}{Plant index. Optional, only for plant or technical parameters}

\item{filepath}{Path to the parameter file}

\item{variety}{Integer. The plant variety to get the parameter from.}

\item{vars}{Vector of variable names for STICS output requirements}
}
\description{
Replace or set an input parameter from a pre-existing STICS input
file. Generally used after calling \code{\link[=import_usm]{import_usm()}}.
}
\details{
The \code{plant} parameter can be either equal to \code{1}, \code{2} for
the associated plant in the case of intercrop, or \code{c(1,2)} for both
Princiapl and associated plants.
\code{\link[=all_out_var]{all_out_var()}} is a helper function that returns all possible
output variables.
}
\note{
\code{set_out_var} is not used by \code{set_param}. To replace the output
variables required from STICS, please directly call \code{set_out_var}.
}
\examples{
\dontrun{
# Replace the interrow distance parameter to 0.01:

library(SticsOnR)
set_param(dirpath = "stics_usm/usm_1", param= "interrang", value= 0.01)

}

}
\seealso{
\code{\link[=import_usm]{import_usm()}}.
}
