% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_exe_utilities.R
\name{set_stics_exe}
\alias{set_stics_exe}
\title{Adding a new Stics model executable in JavaStics}
\usage{
set_stics_exe(javastics, stics_exe, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{javastics}{JavaStics installation root folder}

\item{stics_exe}{Stics executable name (identifier) or executable path}

\item{overwrite}{Boolean. If \code{stics_exe} is an executable path and an
executable with the same name already exist in the bin, overwrite it
if \code{TRUE}, or return an error if \code{FALSE} default.}

\item{verbose}{Logical value (optional), \code{TRUE} to display messages,
\code{FALSE} otherwise.}
}
\value{
Nothing. Update the "preference.xml" file in the config of JavaStics.
}
\description{
Add a new executable of the Stics model to use in JavaStics,
and set it as the one to use by default. Please refer to \code{select_stics_exe()}
to change the stics executable used by JavaStics, and \code{list_stics_exe()}
to list all available executables.
}
\details{
\code{stics_exe} may be :
\enumerate{
\item a model name pointing to a stics executable as done in JavaStics,
e.g. "modulostics" for \code{stics_module.exe};
\item a stics executable file available from the bin folder in JavaStics,
e.g. "stics_module.exe";
\item a path to a stics executable file,
eg. "C:/Users/username/Desktop/stics.exe"
The function will test those hypothesis in the same order they are presented
in the paragraph above. Note that the stics executable will be copied in the
bin folder of JavaStics in the third case, and named after the executable
name and the user's OS, e.g. "stics_intercrop_win" for ane executable called
"stics_intercrop.exe" in a windows OS.
If the file already exists in the bin, the function will return an error
\code{overwrite= FALSE} or will replace the executable if \code{overwrite= TRUE}.
If the name already exist, it will check if the executable is the same.
If it is, the same name is used, if not, a new name with an incremented index
is given, e.g. stics_intercrop_win_2.
In any case, the function will inform the user of which stics is being used
to avoid any issue.
}
}
\note{
"stics_modulo", "sticsmodulo" and "modulostics" are synonyms for the
standard STICS executable.
}
\examples{
\dontrun{
# Using model name:
set_stics_exe("/path/to/JavaSTICS/dir", "modulostics")
# Using model executable:
set_stics_exe("/path/to/JavaSTICS/dir", "stics_modulo.exe")
# Using path to add a new executable:
set_stics_exe("/path/to/JavaSTICS/dir", "path/to/stics.exe")
}

}
\keyword{internal}
