% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stati_stics.R
\name{stati_stics}
\alias{stati_stics}
\title{Compute simulated/observed statistics}
\usage{
stati_stics(..., obs_name = NULL)
}
\arguments{
\item{...}{Either a folder path or the output from \code{\link[=eval_output]{eval_output()}}. If several objects are
detected, make a comparison between them.}

\item{obs_name}{A vector of observation file name(s). Optionnal, required only if \code{...} is a link.
Must have the form \code{c(Dominant,Dominated)} for mixed crops. See details.}
}
\value{
A data.frame with statistics for each simulation. The data.frame has a
\code{description} attribute that gives the description of the column names
}
\description{
Compute statistics for evaluation of the STICS outputs against observations. This function can
be used for one USM or to compare outputs from different model versions or parameter values.
}
\details{
If \code{obs_name} is not provided, the function tries to guess it using the built-in algorithm
from \code{\link[=read_obs]{read_obs()}}. See respective documentation for more details.
}
\note{
Because this function has the purpose to assess model quality, all statistics
are computed on dates were observations are present only. So the simulation mean
is only the mean on dates with observations, not the overall simulation mean.
}
\examples{
\dontrun{
library(sticRs)
# Exemple 1:
tests= stati_stics("dummy/path/simulation",
                   obs_name = c("Wheat.obs","Pea.obs"))
attr(tests,which = "description")

# Exemple 2, stats for a comparison between simulations:
stati_stics(Simulation_1,Simulation_2,
            obs_name = c("Wheat.obs","Pea.obs"))
# Equivalent to:
stati_stics("dummy/path/simulation_1","dummy/path/simulation_2",
            obs_name = c("Wheat.obs","Pea.obs"))

}

}
\seealso{
This function is largely inspired from the \code{evaluate()} function from the
\code{SticsEvalR} package
}
