% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_eval.R
\name{stics_eval}
\alias{stics_eval}
\title{Evaluate STICS version}
\usage{
stics_eval(
  dir.orig = NULL,
  dir.targ = getwd(),
  stics,
  Parameter = NULL,
  Plant = 1,
  obs_name = NULL,
  Out_var = c("lai(n)", "masec(n)", "hauteur"),
  plot_it = T,
  Parallel = T,
  mixed = NULL,
  Title = NULL,
  Erase = TRUE
)
}
\arguments{
\item{dir.orig}{Path to the directory from which to copy the simulation files. If
\code{NULL} (the default), uses the package dummy USM.}

\item{dir.targ}{Path to the target directory for evaluation. Created if missing.}

\item{stics}{STICS executable path named list (see details).}

\item{Parameter}{STICS input parameter named list (see details).}

\item{Plant}{Integer value of the plant on which to set the parameter if STICS
is run on mixed crop (1= Principal, 2= Associated)}

\item{obs_name}{A vector of observation file name(s). It must have the form
\code{c(Dominant,Dominated)} for mixed crops.
See \code{\link[=read_obs]{read_obs()}} \code{filename} parameter for more details.}

\item{Out_var}{The variables needed as output}

\item{plot_it}{Boolean. Do the plot has to be pinted ?}

\item{Parallel}{Are the simulations to be executed in parallel ?}

\item{mixed}{(optional) Is the simulation made on mixed species (boolean)}

\item{Title}{A title for the evaluation. This will be used on the ggplot object.}

\item{Erase}{Should the simulations data be erased upon import (see details)?}
}
\value{
A list of three objects:
\describe{
\item{outputs}{A list of \code{data.frame} objects corresponding
to the simulation for each value in the \code{stics} or \code{Parameter} arguments. The
data.frame is made by a call to \code{\link[=eval_output]{eval_output()}}.}
\item{gg_object}{A summary plot of the simulations outputs returned as a ggplot object.
Possibly a comparison between simulations if several values are given to the \code{stics}
or \code{Parameter} arguments.}
\item{stats}{A \code{data.frame} of a set of summary statistics for each simulation, computed
by a call to \code{\link[=stati_stics]{stati_stics()}}.}
}
}
\description{
Wrapper function to run STICS simulations and evaluate
outputs using observations and/or other model versions.
}
\details{
The function evaluate STICS outputs either along different model versions \strong{OR}
parameter values, not both at the same time. The method is automatically chosen using the
\code{stics} and \code{Parameter} length. The parameter with a length > 1 will be evaluated.
The names of the \code{stics} or the \code{Parameter} list are used for reference in the outputs
(data.frames and plots). The names are mandatory for \code{Parameter}, but are optionnal for
\code{stics}. If no names are provided for \code{stics}, the function give a dummy name for each
model evaluation.
The format of both \code{stics} or \code{Parameter} parameters is the same: a named list of either
STICS executable path or a named list of parameters value(s).
Please set the \code{Parameter} argument to \code{NULL} (the default) for no parameter changes.
The function run STICS executable in parrallel using all cores from the machine but one.
}
\examples{
\dontrun{
library(sticRs)
# Evaluating a change in a parameter:

Eval_parameter=
 stics_eval(dir.orig = "0-DATA/dummy/IC_Wheat_Pea",
            dir.targ = "2-Simulations/Parameterization",
            stics = list(EquivDens= "0-DATA/stics_executable/EquDens/stics.exe"),
            Parameter =  list(interrang= c(0.05, 0.25)),
            obs_name =  c("6_IC_Wheat_N02.obs","6_IC_Pea_N02.obs"),
            Out_var = c("raint", "trg(n)", "rsoleil", "lai(n)", "masec(n)",
            "hauteur", "cumraint", "fapar", "eai"),
            Title = "Wheat-Pea Auzeville 2005-2006 N0", plot_it = T)


# Evaluating a change in the executable:
Eval_stics=
stics_eval(dir.orig = "0-DATA/dummy/IC_Wheat_Pea",
           dir.targ = "2-Simulations/Parameterization",
           stics = list(EquivDens= "0-DATA/stics_executable/EquDens/stics.exe"),
           obs_name =  c("6_IC_Wheat_N02.obs","6_IC_Pea_N02.obs"),
           Out_var = c("raint", "trg(n)", "rsoleil", "lai(n)", "masec(n)",
           "hauteur", "cumraint", "fapar", "eai"),
           Title = "Wheat-Pea Auzeville 2005-2006 N0", plot_it = T)
}
}
\seealso{
\code{\link[=sensitive_stics]{sensitive_stics()}} to evaluate STICS sensitivity to parameter(s), and other
functions used under the hood: \code{\link[=eval_output]{eval_output()}}, \code{\link[=import_usm]{import_usm()}},
\code{\link[=run_stics]{run_stics()}}, \code{\link[=stati_stics]{stati_stics()}}, and \code{\link[=set_out_var]{set_out_var()}}.
}
