% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimi_stics.R
\name{stics_eval_no_copy}
\alias{stics_eval_no_copy}
\title{Make a simulation}
\usage{
stics_eval_no_copy(USM_path, param, obs_name, Plant)
}
\arguments{
\item{USM_path}{The path to the USM folder}

\item{param}{The parameter values}

\item{obs_name}{The observation file names}

\item{Plant}{The plant (\emph{i.e.} Principal or associated) for which the parameters
will be set (only for plant or technical parameters in mixed crop simulations)
Set to \code{NULL} if using STICS in sole crop#'}
}
\value{
The output of \code{\link[=eval_output]{eval_output()}}
}
\description{
Make a STICS simulation such as for using \code{\link[=stics_eval]{stics_eval()}} but without
importing the files, and without making plots. This function is parallelized over USMs.
}
