% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_wrapper.R
\name{stics_wrapper_options}
\alias{stics_wrapper_options}
\title{Getting a stics_wrapper options list with initialized fields}
\usage{
stics_wrapper_options(stics_path, data_dir, ...)
}
\arguments{
\item{stics_path}{Path of the Stics binary executable file (delivered with
JavaStics interface)}

\item{data_dir}{Path(s) of the situation(s) input files directorie(s)
or the root path of the situation(s) input files directorie(s)}

\item{...Add}{further arguments to the options}
}
\value{
A list containing Stics model stics_wrapper options
}
\description{
This function returns a default options list
}
