% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_wrapper.R
\name{stics_wrapper_options}
\alias{stics_wrapper_options}
\title{Getting/setting a stics_wrapper options list with initialized fields,
or validating an existing list}
\usage{
stics_wrapper_options(
  stics_path = NULL,
  data_dir = NULL,
  in_options = NULL,
  ...
)
}
\arguments{
\item{stics_path}{Path of the Stics binary executable file (delivered with
JavaStics interface)}

\item{data_dir}{Path(s) of the situation(s) input files directorie(s)
or the root path of the situation(s) input files directorie(s)}

\item{in_options}{An existing options list, for updating elements}

\item{...}{Add further arguments set the options list values}
}
\value{
A list containing Stics model stics_wrapper options
}
\description{
This function returns a default options list, or allow to set
list elements values or validate the input options list elements values
(only the mandatories ones)
}
\examples{

\dontrun{
# Getting simulations options and defaults values for the stics_wrapper function

stics_wrapper_options()

#> $stics_path
#> "/path/to/javastics/bin/stics_modulo"
#>
#> $data_dir
#> "/path/to/usms/subdirs/root"
#>
#> $parallel
#> [1] FALSE
#>
#> $cores
#> [1] NA
#>
#> $time_display
#> [1] FALSE
#>
#> $warning_display
#> [1] TRUE

# Setting mandatory simulations options
sim_options <- stics_wrapper_options(stics_path = stics_path, data_dir = data_path)

# Setting other options using option name as a function argument
# Example for activating parallel simulations
sim_options <- stics_wrapper_options(stics_path = stics_path,
data_dir = data_path, parallel = TRUE)

#> $stics_path
#> "/path/to/javastics/bin/stics_modulo"
#>
#> $data_dir
#> "/path/to/usms/subdirs/root"
#>
#> $parallel
#> [1] TRUE
#>
#> $cores
#> [1] NA
#>
#> $time_display
#> [1] FALSE
#>
#> $warning_display
#> [1] TRUE

}

}
